chore: split platform registrations into platform specific files
Fixes #1337, also fixes #1338 and closes #1339
