+++ src/ReactiveUI.XamForms/Registrations.cs
@@ -1,25 +1,11 @@
-﻿// Licensed to the .NET Foundation under one or more agreements.
+// Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MS-PL license.
 // See the LICENSE file in the project root for more information.
 
 using System;
-using System.Collections.Generic;
-using System.ComponentModel;
-using System.Diagnostics;
-using System.Linq;
-using System.Text;
-using System.Windows;
-using ReactiveUI;
-using System.Reactive.Concurrency;
-using Splat;
 
 namespace ReactiveUI.XamForms
 {
-    /// <summary>
-    /// Ignore me. This class is a secret handshake between RxUI and RxUI.Xaml
-    /// in order to register certain classes on startup that would be difficult
-    /// to register otherwise.
-    /// </summary>
     public class Registrations : IWantsToRegisterStuff
     {
         public void Register(Action<Func<object>, Type> registerFunction)
