+++ src/ReactiveUI/Platforms/uap10.0/PlatformRegistrations.cs
@@ -1,0 +1,24 @@
+// Licensed to the .NET Foundation under one or more agreements. 
+// The .NET Foundation licenses this file to you under the MS-PL license. 
+// See the LICENSE file in the project root for more information. 
+
+using System;
+using System.Reactive.Concurrency;
+
+namespace ReactiveUI
+{
+    public class PlatformRegistrations : IWantsToRegisterStuff
+    {
+        public void Register(Action<Func<object>, Type> registerFunction)
+        {
+            registerFunction(() => new PlatformOperations(), typeof(IPlatformOperations));
+            registerFunction(() => new ActivationForViewFetcher(), typeof(IActivationForViewFetcher));
+            registerFunction(() => new DependencyObjectObservableForProperty(), typeof(ICreatesObservableForProperty));
+            registerFunction(() => new BooleanToVisibilityTypeConverter(), typeof(IBindingTypeConverter));
+            registerFunction(() => new AutoDataTemplateBindingHook(), typeof(IPropertyBindingHook));
+            RxApp.TaskpoolScheduler = TaskPoolScheduler.Default;            
+            RxApp.MainThreadScheduler = new WaitForDispatcherScheduler(() => CoreDispatcherScheduler.Current);
+            registerFunction(() => new WinRTAppDataDriver(), typeof(ISuspensionDriver));
+        }
+    }
+}
