+++ src/ReactiveUI/Registrations.cs
@@ -3,17 +3,13 @@
 // See the LICENSE file in the project root for more information.
 
 using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace ReactiveUI
 {
     public class Registrations : IWantsToRegisterStuff
     {
         public void Register(Action<Func<object>, Type> registerFunction)
-        {            
+        {
             registerFunction(() => new INPCObservableForProperty(), typeof(ICreatesObservableForProperty));
             registerFunction(() => new IROObservableForProperty(), typeof(ICreatesObservableForProperty));
             registerFunction(() => new POCOObservableForProperty(), typeof(ICreatesObservableForProperty));
